+++ modules/flexibleSearch/exec/src/sap/commerce/toolset/flexibleSearch/exec/FlexibleSearchExecClient.kt
@@ -52,7 +52,7 @@ class FlexibleSearchExecClient(
             .map { BasicNameValuePair(it.key, it.value) }
 
         val response = HacHttpClient.getInstance(project)
-            .post(actionUrl, params, true, context.settings.timeout, connection, null)
+            .post(actionUrl, params, true, context.timeout, connection, null)
         val statusLine = response.statusLine
         val statusCode = statusLine.statusCode