+++ modules/impex/ui/src/sap/commerce/toolset/impex/editor/ImpExInEditorParametersView.kt
@@ -73,13 +73,8 @@ class ImpExInEditorParametersView(private val project: Project, private val coro
         }
 
         return panel {
-            notificationPanel()
-
-            if (virtualParameters.isEmpty()) {
-                notResultsPanel()
-            } else {
-                parametersPanel(virtualParameters, fileEditor)
-            }
+            if (virtualParameters.isEmpty()) notResultsPanel()
+            else parametersPanel(virtualParameters, fileEditor)
         }
             .apply {
                 border = JBUI.Borders.empty(5, 16, 10, 16)
@@ -92,24 +87,6 @@ class ImpExInEditorParametersView(private val project: Project, private val coro
                 isFocusCycleRoot = true
             }
     }
-
-    private fun Panel.notificationPanel() = panel {
-        row {
-            cell(
-                InlineBanner(
-                    """
-                            <html><body style='width: 100%'>
-                            <p>This feature is experimental and may be unstable. Use with caution.</p>
-                            <p>Submit issues or suggestions to project's GitHub <a href="https://github.com/epam/sap-commerce-intellij-idea-plugin/issues/new">repository</a>.</p>
-                            </body></html>
-                        """.trimIndent(),
-                    EditorNotificationPanel.Status.Promo
-                )
-            )
-                .align(Align.FILL)
-                .resizableColumn()
-        }.topGap(TopGap.SMALL)
-    }.customize(UnscaledGaps(16, 16, 16, 16))
 
     private fun Panel.notResultsPanel() = panel {
         row {
@@ -164,7 +141,7 @@ class ImpExInEditorParametersView(private val project: Project, private val coro
                         .filter { (key, _) -> key.element?.isEquivalentTo(it) ?: false }
                         .map { (pointer, virtualParameter) -> pointer to virtualParameter }
                         .firstOrNull()
-                        ?.takeIf { (_, virtualParameter) -> it.text == virtualParameter.name}
+                        ?.takeIf { (_, virtualParameter) -> it.text == virtualParameter.name }
                         ?: (SmartPointerManager.createPointer(it) to ImpExVirtualParameter.of(it))
                 }
                 ?: emptyMap()